api/requirements.txt
fastapi
uvicorn
python-docx
lxml
firebase-admin
python-multipart
python-dotenv
